RING-28615 Split functional tests into GET/PUT/DELETE suites
