+++ tests/functional/test_get.js
@@ -26,118 +26,14 @@ function getDefaultClient() {
     return client;
 }
 
-mocha.describe('Hyperdrive Client Single endpoint suite', function () {
+mocha.describe('Hyperdrive Client GET', function () {
     // Clean all HTTP mocks before starting the test
     mocha.beforeEach(nock.cleanAll);
 
     // Verify all mocks have been used - no garbage
-    mocha.afterEach(function () {
-        assert.ok(nock.isDone);
-    });
+    mocha.afterEach(function () { assert.ok(nock.isDone); });
 
-    mocha.describe('Internals', function () {
-        mocha.it('Socket error handling', function (done) {
-            const hdClient = getDefaultClient();
-            const [ip, port] = hdClient.options.policy.locations[0].split(':');
-            const opts = hdClient._getCommonStoreRequestOptions(
-                ip, Number(port), 'test_key');
-            opts.method = 'GET';
-            opts.path = '/jesuisunemiteenpullover';
-            const noLog = { error() {} };
-            const expectedErrorMessage = 'something awful happened';
-
-            nock(`http://${hdClient.options.policy.locations[0]}`)
-                 .get(opts.path)
-                .replyWithError(expectedErrorMessage);
-
-            const opContext = hdclient.httpUtils.makeOperationContext(
-                { nDataParts: 1, nCodingParts: 0, nChunks: 1 });
-            const reqContext = {
-                opContext,
-                chunkId: 0,
-                fragmentId: 0,
-            };
-
-            hdclient.httpUtils.newRequest(
-                opts, noLog, reqContext, 0, opCtx => {
-                    assert.strictEqual(opCtx.status[0].nError, 1);
-                    assert.ok(opCtx.status[0].statuses[0].error);
-                    const returnedError = opCtx.status[0].statuses[0].error;
-                    assert.strictEqual(expectedErrorMessage,
-                                       returnedError.message);
-                    done();
-                }).end();
-        });
-    });
-
-    mocha.describe('DELETE', function () {
-        mocha.it('Existing key', function (done) {
-            const hdClient = getDefaultClient();
-            const mocks = [
-                { statusCode: 200 },
-            ];
-            const { rawKey } = hdmock.mockDELETE(
-                hdClient.options, 'bestObjEver', mocks);
-            hdClient.delete(rawKey, '1', err => {
-                done(err);
-            });
-        });
-
-        mocha.it('Not found key', function (done) {
-            const hdClient = getDefaultClient();
-            const mocks = [
-                { statusCode: 404 },
-            ];
-            const { rawKey } = hdmock.mockDELETE(
-                hdClient.options, 'bestObjEver', mocks);
-            hdClient.delete(rawKey, '1', err => {
-                assert.strictEqual(err.infos.status, mocks[0].statusCode);
-                done();
-            });
-        });
-
-        mocha.it('Server error', function (done) {
-            const hdClient = getDefaultClient();
-            const mocks = [
-                { statusCode: 500 },
-            ];
-            const { rawKey } = hdmock.mockDELETE(
-                hdClient.options, 'bestObjEver', mocks);
-            hdClient.delete(rawKey, '1', err => {
-                assert.strictEqual(err.infos.status, mocks[0].statusCode);
-                done();
-            });
-        });
-
-        mocha.it('Bad key', function (done) {
-            const hdClient = getDefaultClient();
-            hdClient.delete('---', '1', err => {
-                if (!(err instanceof BadKeyError)) {
-                    throw err;
-                }
-                done();
-            });
-        });
-
-        mocha.it('Timeout', function (done) {
-            const hdClient = getDefaultClient();
-            const mocks = [
-                {
-                    statusCode: 200,
-                    timeoutMs: hdClient.options.requestTimeoutMs + 10,
-                },
-            ];
-            const { rawKey } = hdmock.mockDELETE(
-                hdClient.options, 'bestObjEver', mocks);
-
-            hdClient.delete(rawKey, '1', err => {
-                assert.ok(err);
-                done();
-            });
-        });
-    });
-
-    mocha.describe('GET', function () {
+    mocha.describe('Single hyperdrive', function () {
         mocha.it('Existing small key', function (done) {
             const hdClient = getDefaultClient();
             const content = 'Je suis une mite en pullover';
@@ -450,153 +346,4 @@ mocha.describe('Hyperdrive Client Single endpoint suite', function () {
                 });
         });
     });
-
-    mocha.describe('PUT', function () {
-        mocha.it('Success small key', function (done) {
-            const hdClient = getDefaultClient();
-            const mocks = [
-                {
-                    statusCode: 200,
-                    payload: 'Je suis une mite en pullover',
-                    contentType: 'data',
-                },
-            ];
-            const keyContext = {
-                objectKey: 'bestObjEver',
-            };
-
-            hdmock.mockPUT(hdClient.options, keyContext, mocks);
-
-            hdClient.put(
-                hdmock.streamString(mocks[0].payload),
-                hdmock.getPayloadLength(mocks[0].payload),
-                keyContext, '1',
-                (err, rawKey) => {
-                    assert.ifError(err);
-                    assert.strictEqual(typeof rawKey, 'string');
-                    const parts = hdclient.keyscheme.deserialize(rawKey);
-
-                    const [endpoint, port] =
-                              hdClient.options.policy.locations[0].split(':');
-                    assert.strictEqual(parts.nDataParts, 1);
-                    assert.strictEqual(parts.nCodingParts, 0);
-                    assert.strictEqual(parts.nChunks, 1);
-                    const fragment = parts.chunks[0].data[0];
-                    assert.strictEqual(fragment.hostname, endpoint);
-                    assert.strictEqual(fragment.port, Number(port));
-                    assert.ok(fragment.key, keyContext.objectKey);
-                    done();
-                });
-        });
-
-        mocha.it('Success larger key (32 KiB)', function (done) {
-            const hdClient = getDefaultClient();
-            /* TODO avoid depending on hardcoded path */
-            const content = fs.createReadStream(
-                'tests/functional/random_payload');
-            const mocks = [
-                {
-                    statusCode: 200,
-                    payload: content,
-                    contentType: 'data',
-                },
-            ];
-            const keyContext = {
-                objectKey: 'bestObjEver',
-            };
-
-            hdmock.mockPUT(hdClient.options, keyContext, mocks);
-
-            hdClient.put(
-                content,
-                hdmock.getPayloadLength(content),
-                keyContext, '1',
-                (err, rawKey) => {
-                    assert.ifError(err);
-                    assert.strictEqual(typeof rawKey, 'string');
-                    const parts = hdclient.keyscheme.deserialize(rawKey);
-
-                    const [endpoint, port] =
-                              hdClient.options.policy.locations[0].split(':');
-                    assert.strictEqual(parts.nDataParts, 1);
-                    assert.strictEqual(parts.nCodingParts, 0);
-                    assert.strictEqual(parts.nChunks, 1);
-                    const fragment = parts.chunks[0].data[0];
-                    assert.strictEqual(fragment.hostname, endpoint);
-                    assert.strictEqual(fragment.port, Number(port));
-                    assert.ok(fragment.key, keyContext.objectKey);
-                    done();
-                });
-        });
-
-        mocha.it('Server error', function (done) {
-            const hdClient = getDefaultClient();
-            const mocks = [
-                {
-                    statusCode: 500,
-                    payload: 'Je suis une mite en pullover',
-                    contentType: 'data',
-                },
-            ];
-            const keyContext = {
-                objectKey: 'bestObjEver',
-            };
-
-            hdmock.mockPUT(hdClient.options, keyContext, mocks);
-
-            hdClient.put(
-                hdmock.streamString(mocks[0].payload),
-                hdmock.getPayloadLength(mocks[0].payload),
-                keyContext, '1',
-                err => {
-                    assert.strictEqual(err.infos.status, mocks[0].statusCode);
-                    done();
-                });
-        });
-
-        mocha.it('Timeout', function (done) {
-            const hdClient = getDefaultClient();
-            const mocks = [
-                {
-                    statusCode: 404,
-                    payload: 'Je suis une mite en pullover',
-                    contentType: 'data',
-                    timeoutMs: hdClient.options.requestTimeoutMs + 10,
-                },
-            ];
-            const keyContext = {
-                objectKey: 'bestObjEver',
-            };
-
-            hdmock.mockPUT(hdClient.options, keyContext, mocks);
-
-            let called = false;
-            hdClient.put(
-                hdmock.streamString(mocks[0].payload),
-                hdmock.getPayloadLength(mocks[0].payload),
-                keyContext, '1',
-                /* PUT is considered successful on timeout as we don't
-                 * whether it really was stored or not
-                 */
-                (err, rawKey) => {
-                    assert.ok(!called);
-                    called = true;
-                    assert.ifError(err);
-                    assert.strictEqual(typeof rawKey, 'string');
-                    const parts = hdclient.keyscheme.deserialize(rawKey);
-
-                    const [endpoint, port] =
-                              hdClient.options.policy.locations[0].split(':');
-                    assert.strictEqual(parts.nDataParts, 1);
-                    assert.strictEqual(parts.nCodingParts, 0);
-                    assert.strictEqual(parts.nChunks, 1);
-                    const fragment = parts.chunks[0].data[0];
-                    assert.strictEqual(fragment.hostname, endpoint);
-                    assert.strictEqual(fragment.port, Number(port));
-                    assert.ok(fragment.key, keyContext.objectKey);
-                    done();
-                });
-        });
-    });
 });
-
